[AXON-45] First approximation of API token auth for JIRA
